Health\Damage type
Изменён тип здоровья\урона на целочисленный

+++ Assets/Code/Scripts/Logic/UnitAttack.cs
@@ -5,18 +5,18 @@ public class UnitAttack : MonoBehaviour
 {
     [SerializeField] private Unit _unit;
 
-    public event Action<float> OnDamageChanged;
-    public float Damage
+    public event Action<int> OnDamageChanged;
+    public int Damage
     {
         get => _damage;
         set
         {
-            _damage = Mathf.Clamp(value, 0f, float.MaxValue);
+            _damage = Mathf.Clamp(value, 0, int.MaxValue);
             OnDamageChanged?.Invoke(_damage);
         }
     }
 
-    private float _damage;
+    private int _damage;
 
     private void Awake()
     {
